requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
click-plugins==1.1.1
colorama==0.3.7
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
etelemetry==0.3.1
executing==2.1.0
filelock==3.16.1
fitz==0.0.1.dev2
frozenlist==1.5.0
future==1.0.0
geoip2==4.8.0
httplib2==0.22.0
idna==3.10
ipwhois==1.3.0
ipykernel==6.29.5
ipython==8.29.0
isodate==0.6.1
jedi==0.19.2
jsbeautifier==1.6.2
jupyter_client==8.6.3
jupyter_core==5.7.2
looseversion==1.3.0
lxml==5.3.0
matplotlib-inline==0.1.7
maxminddb==2.6.2
maxminddb-geolite2==2018.703
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
peepdf==0.4.2
pexpect==4.9.0
Pillow==3.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
prov==2.0.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pydot==3.0.2
pygeoip==0.3.2
Pygments==2.18.0
PyMuPDF==1.24.13
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-geoip==1.2
python-geoip-geolite2==2015.303
python-geoip-python3==1.3
python-whois==0.9.4
pythonaes==1.0
pytz==2024.2
pyxnat==1.6.2
pyzmq==26.2.0
rdflib==6.3.2
requests==2.32.3
requests-file==2.1.0
scipy==1.14.1
setuptools==75.5.0
shodan==1.31.0
simplejson==3.19.3
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tldextract==5.1.3
tornado==6.4.1
traitlets==5.14.3
traits==6.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
wheel==0.45.0
whois==1.20240129.2
XlsxWriter==3.2.0
yarl==1.17.1
